Mask secrets - tests, refactoring, null check
Extracted ExecutionContext inner class to package-private class CommandRunnerExecutionContext. Was too big and deserved a file on its own.
Moved mask method to ParameterMap and added unit tests.
